+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.4.0] - 2020-10-10
+
+* Fix invalid command.
+* Add option to disable `--delete-conflicting-outputs`.
+
 ## [0.3.0] - 2020-10-07
 
 * Improve argument handling with `args` package.